Update benchmark/profile functions
